Update OPCUADataSelectionConfiguration_1-0.md
Hi Tina, this is just a small change to make correct description of the name property for OPC UA EDS adapter. The current description is valid for the released standalone version of the adapter (1.1) but the EDS required name field to be specified. Please feel free to change the wording and reach out.

Thank you,
Zdenek